refactor Main endpoints Else cosas
